disable make without explicit 1st-arg type, like make(2, 3)
Having the shortcut `make(2, 3)` for `make(Array, 2, 3)`, which mirrors
`rand(2, 3)`, is convenient, but the form `make(Int, 2, 3)` leads to
ambiguities for types which take 2 integer args, e.g.
`make(MyType, 2, 3)`. Let's favor simplicity for custom-defined types.

Moreover,  the terminology "make" makes sense when the first arg
refers to the type to be "made", e.g. in `make(Int, 2, 3)` we didn't
make an `Int`, but an array, so the meaning of "make" was lost.