allow non-type as first argument of make(x...) (#11)
Till now, `make(x)` was defined to create an explicit distribution wrapping
a possibly implicit one (`x`). Beside this case, the first argument of
`make(x1, xs...)` had to be a type (typically a supertype of the `gentype` of
the resulting "make value").

But there is no reason to prevent allowing a non-type object as the first
argument; in this case though, this non-type object `x1` has to be passed as
an argument to `Make` (i.e. the resulting `Make` object contains `x1` as
its first element), because keeping only its type would be lossy.